backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.1.0
channels-redis==4.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
redis==5.0.6

# RAG System Dependencies
openai>=1.6.1
torch>=2.0.0
transformers>=4.36.2
huggingface-hub>=0.30.0
sentence-transformers>=2.2.2
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
PyPDF2>=3.0.0

# AI Agents Dependencies
anthropic==0.18.1
langchain==0.1.9
langchain-openai==0.0.2
chromadb==0.4.22
tiktoken==0.5.2
faiss-cpu==1.7.4

# Async and Memory
aioredis==2.0.1
celery==5.3.4

# RAG Evaluation Dependencies
deepeval==0.21.73
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0

# Additional AI Dependencies
typing-extensions==4.8.0 